feat(inverter): Refactor inverter package structure and update imports
