+++ src/main/java/com/youssef/GridPulse/domain/device/service/DeviceService.java
@@ -15,8 +15,8 @@ import com.youssef.GridPulse.domain.fleet.repository.FleetRepository;
 import com.youssef.GridPulse.domain.identity.user.Role;
 import com.youssef.GridPulse.domain.identity.user.entity.User;
 import com.youssef.GridPulse.domain.identity.user.repository.UserRepository;
-import com.youssef.GridPulse.domain.inverter.entity.Inverter;
-import com.youssef.GridPulse.domain.inverter.repository.InverterRepository;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.Inverter;
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterRepository;
 import com.youssef.GridPulse.domain.meter.entity.Meter;
 import com.youssef.GridPulse.domain.meter.repository.MeterRepository;
 import jakarta.persistence.EntityNotFoundException;
