+++ src/main/java/com/youssef/GridPulse/domain/inverter/inverter/dto/InverterInput.java
@@ -1,4 +1,4 @@
-package com.youssef.GridPulse.domain.inverter.dto;
+package com.youssef.GridPulse.domain.inverter.inverter.dto;
 
 public record InverterInput(
         String name,
