+++ src/main/java/com/youssef/GridPulse/domain/inverter/inverter/service/InverterService.java
@@ -1,12 +1,12 @@
-package com.youssef.GridPulse.domain.inverter.service;
+package com.youssef.GridPulse.domain.inverter.inverter.service;
 
 import com.youssef.GridPulse.common.base.BaseService;
-import com.youssef.GridPulse.domain.inverter.dto.InverterInput;
-import com.youssef.GridPulse.domain.inverter.entity.Inverter;
-import com.youssef.GridPulse.domain.inverter.entity.InverterHistory;
-import com.youssef.GridPulse.domain.inverter.mapper.InverterMapper;
-import com.youssef.GridPulse.domain.inverter.repository.InverterHistoryRepository;
-import com.youssef.GridPulse.domain.inverter.repository.InverterRepository;
+import com.youssef.GridPulse.domain.inverter.inverter.dto.InverterInput;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.Inverter;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.InverterHistory;
+import com.youssef.GridPulse.domain.inverter.inverter.mapper.InverterMapper;
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterHistoryRepository;
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterRepository;
 
 import org.springframework.stereotype.Service;
 import java.util.UUID;
