+++ src/test/java/com/youssef/GridPulse/domain/base/BaseResolverTest.java
@@ -3,7 +3,7 @@ package com.youssef.GridPulse.domain.base;
 import com.youssef.GridPulse.common.base.BaseEntity;
 import com.youssef.GridPulse.common.base.BaseHistoryEntity;
 import com.youssef.GridPulse.common.base.BaseService;
-import com.youssef.GridPulse.domain.inverter.entity.Inverter;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.Inverter;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.test.tester.GraphQlTester;
