+++ src/test/java/com/youssef/GridPulse/domain/inverter/repository/InverterRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.youssef.GridPulse.domain.inverter.repository;
 
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterRepository;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
