+++ src/test/java/com/youssef/GridPulse/domain/inverter/resolver/InverterResolverTests.java
@@ -1,10 +1,11 @@
 package com.youssef.GridPulse.domain.inverter.resolver;
 
 import com.youssef.GridPulse.domain.base.BaseResolverTest;
-import com.youssef.GridPulse.domain.inverter.dto.InverterInput;
-import com.youssef.GridPulse.domain.inverter.entity.Inverter;
-import com.youssef.GridPulse.domain.inverter.entity.InverterHistory;
-import com.youssef.GridPulse.domain.inverter.service.InverterService;
+import com.youssef.GridPulse.domain.inverter.inverter.dto.InverterInput;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.Inverter;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.InverterHistory;
+import com.youssef.GridPulse.domain.inverter.inverter.resolver.InverterResolver;
+import com.youssef.GridPulse.domain.inverter.inverter.service.InverterService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.graphql.GraphQlTest;
