+++ src/test/java/com/youssef/GridPulse/domain/inverter/service/InverterServiceTest.java
@@ -1,12 +1,13 @@
 package com.youssef.GridPulse.domain.inverter.service;
 
 import com.youssef.GridPulse.domain.base.BaseServiceTest;
-import com.youssef.GridPulse.domain.inverter.dto.InverterInput;
-import com.youssef.GridPulse.domain.inverter.entity.Inverter;
-import com.youssef.GridPulse.domain.inverter.entity.InverterHistory;
+import com.youssef.GridPulse.domain.inverter.inverter.dto.InverterInput;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.Inverter;
+import com.youssef.GridPulse.domain.inverter.inverter.entity.InverterHistory;
+import com.youssef.GridPulse.domain.inverter.inverter.service.InverterService;
 import com.youssef.GridPulse.domain.inverter.mapper.InverterMapperImpl;
-import com.youssef.GridPulse.domain.inverter.repository.InverterHistoryRepository;
-import com.youssef.GridPulse.domain.inverter.repository.InverterRepository;
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterHistoryRepository;
+import com.youssef.GridPulse.domain.inverter.inverter.repository.InverterRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.mockito.Mock;
